backend-doc-images/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.9
pydantic==2.9.2
pymupdf==1.24.10
Pillow==10.4.0
numpy==2.1.3
opencv-python-headless==4.10.0.84
pytesseract==0.3.13